Don't remove papers from the queue during breaks

+++ pcs-chair-site/app/controllers/venues_controller.rb
@@ -72,7 +72,7 @@ class VenuesController < ApplicationController
     respond_to do |format|
       if @venue.update_attributes(params[:venue])
         qpaper = @venue.papers.find_by_pcs_paper_id(@venue.paper_pcs_id)
-        if qpaper != nil
+        if qpaper != nil and !@venue.breaktime
           qpaper.destroy
         end
